+++ android/app/build.gradle
@@ -38,6 +38,9 @@ android {
             // TODO: Add your own signing config for the release build.
             // Signing with the debug keys for now, so `flutter run --release` works.
             signingConfig = signingConfigs.debug
+            //signingConfig = signingConfigs.release
+            shrinkResources false
+            minifyEnabled false
         }
     }
 }
